it works! yay it's round trip egraph
