+++ src/util.rs
@@ -272,6 +272,7 @@ impl Run {
             RunType::RvsdgRoundTrip,
             RunType::PegConversion,
             RunType::CfgRoundTrip,
+            RunType::RvsdgOptimize,
         ] {
             let default = Run {
                 test_type,
